fix: first commit modifications requested
